refactor: Casts, Observers and Traits
