+++ app/Casts/MassCast.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Models\Casts;
+namespace App\Casts;
 
-use App\Support\Units\Fuel;
+use App\Support\Units\Mass;
 use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 use Illuminate\Database\Eloquent\Model;
 use PhpUnitsOfMeasure\Exception\NonNumericValue;
 use PhpUnitsOfMeasure\Exception\NonStringUnitName;
 
-class FuelCast implements CastsAttributes
+class MassCast implements CastsAttributes
 {
     /**
      * Transform the attribute from the underlying model values.
@@ -19,12 +19,12 @@ class FuelCast implements CastsAttributes
      */
     public function get($model, string $key, $value, array $attributes)
     {
-        if ($value instanceof Fuel) {
+        if ($value instanceof Mass) {
             return $value;
         }
 
         try {
-            return Fuel::make($value, config('phpvms.internal_units.fuel'));
+            return Mass::make($value, config('phpvms.internal_units.mass'));
         } catch (NonNumericValue $e) {
         } catch (NonStringUnitName $e) {
             return $value;
@@ -42,8 +42,8 @@ class FuelCast implements CastsAttributes
      */
     public function set($model, string $key, $value, array $attributes)
     {
-        if ($value instanceof Fuel) {
-            return $value->toUnit(config('phpvms.internal_units.fuel'));
+        if ($value instanceof Mass) {
+            return $value->toUnit(config('phpvms.internal_units.mass'));
         }
 
         return $value;
