+++ app/Models/Expense.php
@@ -2,8 +2,8 @@
 
 namespace App\Models;
 
+use App\Casts\CommaDelimitedCast;
 use App\Contracts\Model;
-use App\Models\Casts\CommaDelimitedCast;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
