+++ app/Models/FlightField.php
@@ -3,6 +3,8 @@
 namespace App\Models;
 
 use App\Contracts\Model;
+use App\Observers\Sluggable;
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Support\Str;
@@ -23,6 +25,7 @@ use Illuminate\Support\Str;
  *
  * @mixin \Eloquent
  */
+#[ObservedBy(Sluggable::class)]
 class FlightField extends Model
 {
     use HasFactory;
