+++ app/Models/FlightFieldValue.php
@@ -3,6 +3,8 @@
 namespace App\Models;
 
 use App\Contracts\Model;
+use App\Observers\Sluggable;
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -34,6 +36,7 @@ use Illuminate\Support\Str;
  *
  * @mixin \Eloquent
  */
+#[ObservedBy(Sluggable::class)]
 class FlightFieldValue extends Model
 {
     use HasFactory;
