+++ app/Models/PirepFieldValue.php
@@ -4,6 +4,8 @@ namespace App\Models;
 
 use App\Contracts\Model;
 use App\Models\Enums\PirepFieldSource;
+use App\Observers\Sluggable;
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Support\Carbon;
@@ -35,6 +37,7 @@ use Illuminate\Support\Str;
  *
  * @mixin \Eloquent
  */
+#[ObservedBy(Sluggable::class)]
 class PirepFieldValue extends Model
 {
     public $table = 'pirep_field_values';
