+++ app/Models/Subfleet.php
@@ -4,8 +4,10 @@ namespace App\Models;
 
 use App\Contracts\Model;
 use App\Models\Enums\AircraftStatus;
-use App\Models\Traits\ExpensableTrait;
-use App\Models\Traits\FilesTrait;
+use App\Observers\SubfleetObserver;
+use App\Traits\ExpensableTrait;
+use App\Traits\FilesTrait;
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -83,6 +85,7 @@ use Spatie\Activitylog\Traits\LogsActivity;
  *
  * @mixin \Eloquent
  */
+#[ObservedBy(SubfleetObserver::class)]
 class Subfleet extends Model
 {
     use ExpensableTrait;
