+++ app/Observers/AircraftObserver.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Observers;
+namespace App\Observers;
 
 use App\Models\Aircraft;
 use App\Support\ICAO;
